API-1109: Move CollectApiError
