+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/ConnectionContext.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Connectivity\Connection\Infrastructure;
 
+use Akeneo\Connectivity\Connection\Application\ConnectionContextInterface;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\Write\Connection;
 use Akeneo\Connectivity\Connection\Domain\Settings\Persistence\Repository\ConnectionRepository;
 use Akeneo\Connectivity\Connection\Domain\WrongCredentialsConnection\Persistence\Query\AreCredentialsValidCombinationQuery;
@@ -14,7 +15,7 @@ use Akeneo\Connectivity\Connection\Domain\WrongCredentialsConnection\Persistence
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class ConnectionContext
+class ConnectionContext implements ConnectionContextInterface
 {
     /** @var AreCredentialsValidCombinationQuery */
     private $areCredentialsValidCombinationQuery;
