+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Application/ErrorManagement/Service/CollectApiErrorSpec.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Connectivity\Connection\Infrastructure\ErrorManagement;
+namespace spec\Akeneo\Connectivity\Connection\Application\ErrorManagement\Service;
 
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Model\Write\BusinessError;
 use Akeneo\Connectivity\Connection\Domain\ErrorManagement\Persistence\Repository\BusinessErrorRepository;
